Change type to size_t

+++ include/whisper.h
@@ -674,7 +674,7 @@ extern "C" {
     struct whisper_state * whisper_get_state_from_context(struct whisper_context * ctx);
     struct whisper_timings * whisper_get_timings_with_state(struct whisper_state * state);
     ggml_backend_t whisper_get_preferred_backend(struct whisper_state * state);
-    ggml_backend_t whisper_get_indexed_backend(struct whisper_state* state, int i);
+    ggml_backend_t whisper_get_indexed_backend(struct whisper_state* state, size_t i);
     size_t whisper_get_backend_count(struct whisper_state* state);
 
 #ifdef __cplusplus
