+++ src/whisper-flat.h
@@ -25,7 +25,7 @@ extern "C" {
     WHISPER_FLAT_API struct whisper_state * whisper_flat_get_state_from_context(struct whisper_context * ctx);
     WHISPER_FLAT_API const char * whisper_flat_get_system_info_json(void);
     WHISPER_FLAT_API ggml_backend_t whisper_flat_get_preferred_backend(struct whisper_state * state);
-    WHISPER_FLAT_API ggml_backend_t whisper_flat_get_indexed_backend(struct whisper_state* state, int i);
+    WHISPER_FLAT_API ggml_backend_t whisper_flat_get_indexed_backend(struct whisper_state* state, size_t i);
     WHISPER_FLAT_API size_t whisper_flat_get_backend_count(struct whisper_state* state);
 
 #ifdef __cplusplus
